Fix #8
Simple copy-and-paste code error resulted in an empty "IsPrerelease" field.
choco.exe rejects this without a clear error message, but instead with
trying to get the same package with a new version, appending ".0" to the old one.
If this fails it states, that the package does not exist.

Tagging new version 1.0.2